drivers/{canbus,hd2001}: message impl for canbus
This CL migrates the underlying implementation of the CANbus communication to a
message-based protocol with a single channel to post requests and where
requests hold the response channel.

Also handle the case where 2 replies are read in one go (when the first reply
indicates the bus needs some more time to answer.)

Fixes go-lsst/fcs-lpc-bench#1.